Allow null intial values for objectives
objective.value is usually undefined when loading a scoresheet.  Missions that depend on undefined values do not try to determine a score.

Allow objective.value to be initialized as null.  [intended for use with RESOURCE-related objectives]

Null objectives are not marked as complete.
A mission that does not calculate a numeric score is not marked as complete.

This implementation allows missions that have objectives used to check a RESOURCE to calculate a score, even if some of the RESOURCE-related objectives have not been set yet.